refactor(backend-service-game): extract buildIoredisModuleOptions
