+++ packages/backend/services/backend-service-game/src/foundation/redis/adapter/ioredis/calculations/buildIoredisModuleOptions.ts
@@ -1,0 +1,21 @@
+import {
+  EnvModule,
+  Environment,
+  EnvironmentService,
+} from '@cornie-js/backend-app-game-env';
+import { PubSubIoredisModuleOptions } from '@cornie-js/backend-game-adapter-ioredis';
+
+export function buildIoredisModuleOptions(): PubSubIoredisModuleOptions {
+  return {
+    imports: [EnvModule],
+    inject: [EnvironmentService],
+    useFactory: (environmentService: EnvironmentService) => {
+      const environment: Environment = environmentService.getEnvironment();
+
+      return {
+        host: environment.pubSubRedisHost,
+        port: environment.pubSubRedisPort,
+      };
+    },
+  };
+}
